pyclient: move test funtion out of zeth module
